Added SVG Jumperless image
I added the image with labeled rows, (both a visible layer on the bottom and a hidden layer on top as a clickable area).
I have no clue what I'm doing though so don't trust anything I did. But at least the image is in there so you can maybe do something with it.

Additional changes:
- extract board background into PNG file
- adjust sizing of ImageViewBoard